ground work for matcher tests (#450)
These are single-refspec matches without anything added to them. We
might just go straight to matchgroups actually as it it make handling
types easier. Let's see.
