+++ git-refspec/src/matcher.rs
@@ -1,4 +1,4 @@
-use crate::Match;
+use crate::Matcher;
 use bstr::BStr;
 use git_hash::{oid, ObjectId};
 
@@ -11,12 +11,30 @@ pub(crate) enum Needle<'a> {
 }
 
 impl<'a> From<&'a BStr> for Needle<'a> {
-    fn from(_v: &'a BStr) -> Self {
+    fn from(v: &'a BStr) -> Self {
+        if v.starts_with(b"refs/") {
+            Needle::FullName(v)
+        } else {
+            todo!()
+        }
+    }
+}
+
+impl<'a> Matcher<'a> {
+    /// For each name in `names`, set the corresponding byte in `matches` to `true` if the corresponding `name` matches the remote side
+    /// instruction (i.e. the left side of a [`fetch`][crate::parse::Operation::Fetch] refspec).
+    /// Note that `name` is expected to be the full name of a reference.
+    pub fn match_remotes<'b>(
+        &self,
+        _names: impl Iterator<Item = Item<'b>> + ExactSizeIterator,
+        _matches: &mut Vec<Match<'a>>,
+    ) {
         todo!()
     }
 }
 
-/// An item to match
+/// An item to match, input to various matching operations.
+#[derive(Debug, Copy, Clone)]
 pub struct Item<'a> {
     /// The full name of the references, like `refs/heads/main`
     pub full_ref_name: &'a BStr,
@@ -43,4 +61,12 @@ impl Match<'_> {
     pub fn local(&self) -> Option<&BStr> {
         self.rhs
     }
+}
+
+/// The result of a match operation.
+#[derive(Default)]
+#[allow(dead_code)]
+pub struct Match<'a> {
+    pub(crate) lhs: Option<&'a bstr::BStr>,
+    pub(crate) rhs: Option<&'a bstr::BStr>,
 }
