+++ git-refspec/src/spec.rs
@@ -134,10 +134,7 @@ mod matcher {
     use crate::{Matcher, RefSpecRef};
 
     impl<'a> RefSpecRef<'a> {
-        /// For each name in `names`, set the corresponding byte in `matches` to `true` if the corresponding `name` matches the remote side
-        /// instruction (i.e. the left side of a [`fetch`][crate::parse::Operation::Fetch] refspec).
-        /// Note that `name` is expected to be the full name of a reference.
-        // TODO: move docs to Matcher, adjust these.
+        /// Create a matcher to perform matching operations against this refspec.
         pub fn to_matcher(&self) -> Matcher<'_> {
             Matcher {
                 op: self.op,
