+++ git-refspec/tests/refspec.rs
@@ -4,4 +4,5 @@ use git_testtools::Result;
 
 mod impls;
 mod matcher;
+mod matching;
 mod parse;
